Change id type from int to Long Add repositories

+++ src/main/java/com/example/AKTours/model/entity/Continent.java
@@ -17,7 +17,7 @@ public class Continent {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int continent_id;
+    private Long continent_id;
 
     @Column(name = "continent_name")
     private String name;
